+++ lib/src/data/custom_builder.dart
@@ -54,7 +54,7 @@ final class CustomBuilder<BeanT extends Object> {
   /// - `selector`: Optional function that allows conditional selection of instances based on specific criteria. Useful for dynamically choosing an instance at runtime based on application context.
   /// - `qualifier`: Optional qualifier name to distinguish between different instances of the same type.
   /// - `canRegister`: Optional function to conditionally register the instance.
-  /// - `required`: Optional set of qualifiers or types that must be registered before creating an instance.
+  /// - `requires`: Optional set of qualifiers or types that must be registered before creating an instance.
   Future<void> asApplication({
     ListDecorator<BeanT> decorators = const [],
     bool canDestroy = true,
@@ -62,7 +62,7 @@ final class CustomBuilder<BeanT extends Object> {
     FutureOr<bool> Function(Object)? selector,
     Object? qualifier,
     FutureOr<bool> Function()? canRegister,
-    Set<Object>? required,
+    Set<Object>? requires,
   }) {
     return DDI.instance.register<BeanT>(
       factory: ApplicationFactory<BeanT>(
@@ -71,7 +71,7 @@ final class CustomBuilder<BeanT extends Object> {
         canDestroy: canDestroy,
         children: children,
         selector: selector,
-        required: required,
+        requires: requires,
       ),
       qualifier: qualifier,
       canRegister: canRegister,
@@ -86,7 +86,7 @@ final class CustomBuilder<BeanT extends Object> {
   /// - `selector`: Optional function that allows conditional selection of instances based on specific criteria. Useful for dynamically choosing an instance at runtime based on application context.
   /// - `qualifier`: Optional qualifier name to distinguish between different instances of the same type.
   /// - `canRegister`: Optional function to conditionally register the instance.
-  /// - `required`: Optional set of qualifiers or types that must be registered before creating an instance.
+  /// - `requires`: Optional set of qualifiers or types that must be registered before creating an instance.
   Future<void> asDependent({
     ListDecorator<BeanT> decorators = const [],
     bool canDestroy = true,
@@ -94,7 +94,7 @@ final class CustomBuilder<BeanT extends Object> {
     FutureOr<bool> Function(Object)? selector,
     Object? qualifier,
     FutureOr<bool> Function()? canRegister,
-    Set<Object>? required,
+    Set<Object>? requires,
   }) {
     return DDI.instance.register<BeanT>(
       factory: DependentFactory<BeanT>(
@@ -103,7 +103,7 @@ final class CustomBuilder<BeanT extends Object> {
         canDestroy: canDestroy,
         children: children,
         selector: selector,
-        required: required,
+        requires: requires,
       ),
       qualifier: qualifier,
       canRegister: canRegister,
@@ -118,7 +118,7 @@ final class CustomBuilder<BeanT extends Object> {
   /// - `selector`: Optional function that allows conditional selection of instances based on specific criteria. Useful for dynamically choosing an instance at runtime based on application context.
   /// - `qualifier`: Optional qualifier name to distinguish between different instances of the same type.
   /// - `canRegister`: Optional function to conditionally register the instance.
-  /// - `required`: Optional set of qualifiers or types that must be registered before creating an instance.
+  /// - `requires`: Optional set of qualifiers or types that must be registered before creating an instance.
   Future<void> asSingleton({
     ListDecorator<BeanT> decorators = const [],
     bool canDestroy = true,
@@ -126,7 +126,7 @@ final class CustomBuilder<BeanT extends Object> {
     FutureOr<bool> Function(Object)? selector,
     Object? qualifier,
     FutureOr<bool> Function()? canRegister,
-    Set<Object>? required,
+    Set<Object>? requires,
   }) {
     return DDI.instance.register<BeanT>(
       factory: SingletonFactory<BeanT>(
@@ -135,7 +135,7 @@ final class CustomBuilder<BeanT extends Object> {
         canDestroy: canDestroy,
         children: children,
         selector: selector,
-        required: required,
+        requires: requires,
       ),
       qualifier: qualifier,
       canRegister: canRegister,